Refactor: Move the unit test runner to the C_Testing API
It's still very bare-bones, but this way it can be used for the evo-luvi runtime, as well.
